Applied Java conventions to IO package (Reader/Writer, Input/OutputStream for ASCII, binary IO). Reduced debug output for non-SPD warning with full covariances. Fixed IOUtil.writeFloat for arrays.

+++ jstk/src/de/fau/cs/jstk/app/GaussEM.java
@@ -141,7 +141,7 @@ public class GaussEM {
 				FrameInputStream r = chunk.getFrameReader();
 				double [] buf = new double [r.getFrameSize()];
 				while (r.read(buf))
-					data.add(new Sample(0, buf));
+					data.add(new Sample((short) 0, buf));
 			}
 			logger.info(data.size() + " samples cached");
